fix: do not decorate `ELLIPSIS` with `<b>` tags (#123)
* fix: do not decorate `ELLIPSIS` with `<b>` tags

Multiple matches are merged with `ELLIPSIS` as a separator.
This reverses the order of operations so that the matches are decorated before
they are merged. This way, the content of `ELLIPSIS` is not decorated with HTML
`<b>` tags.

* fix: use correct module for Iterable

Collections Abstract Base Classes are deprecated in `collections` since
Python 3.3. They will be moved to `collections.abc` after Python 3.8.
https://docs.python.org/3.8/library/collections.html